# remove all_color_aware optimization in benchmark_xl